Move SSL card into a separate component
